+++ local-data-platform/real_world_use_cases/nyc_yellow_taxi_dataset/monthly_reporting.md
@@ -1,0 +1,23 @@
+'''
+New York Taxi and Limousine Commission
+TLC Trip Record Data
+
+Yellow and green taxi trip records include fields capturing pick-up and drop-off dates/times,
+pick-up and drop-off locations, trip distances, itemized fares, rate types, payment types,
+and driver-reported passenger counts.
+
+For-Hire Vehicle (“FHV”) trip records include fields capturing the dispatching base license number and the pick-up date,
+time, and taxi zone location ID (shape file below).
+
+All files will be stored in the PARQUET format.
+Trip data will be published monthly (with two months delay) instead of bi-annually.
+HVFHV files will now include 17 more columns (please see High Volume FHV Trips Dictionary for details).
+Additional columns will be added to the old files as well.
+'''
+
+
+
+
+
+
+
